old-resources/app/requirements.txt
streamlit
google-generativeai
langgraph
langchain
langchain-google-genai
langchain-pinecone
pinecone-client